[Flang] Move genMinMaxlocReductionLoop to Transforms/Utils.cpp
This is one option for attempting to move genMinMaxlocReductionLoop to a better
location. It moves it into Transforms and makes HLFIRTranforms depend upon
FIRTransforms.
